+++ march_hardware/include/march_hardware/Joint.h
@@ -58,8 +58,8 @@ public:
   int16_t getTorque();
   int32_t getAngleIUAbsolute();
   int32_t getAngleIUIncremental();
-  int32_t getVelocityIUAbsolute();
-  int32_t getVelocityIUIncremental();
+  float getVelocityIUAbsolute();
+  float getVelocityIUIncremental();
   float getTemperature();
   IMotionCubeState getIMotionCubeState();
